add Chairmarks benchmark + JS parity fixes
bench/
  Project.toml: separate env with Chairmarks
  bench.jl:     mirrors martini/bench.js (init, create_tile, mesh(30), sweep 0..20)
  RESULTS.md:   side-by-side comparison with martini.js

src/Martini.jl: _update_errors! now does Float64 intermediates and truncates
on store, mirroring how JS's Math.max(Float32Array_read, F64) behaves.

test/util.jl: mapbox_terrain_to_grid decodes in Float64 and truncates on
store; previously F32 throughout, which produced ~1 ULP diffs that flipped
subdivision decisions at low max_error.

Net effect: byte-identical mesh output vs JS at every threshold tested
(10, 30, 100, 500, 1000). The existing max_error=500 parity test still passes.

Headline timings (fuji.png 513-grid):
  Mesher(513)      JS 19.7 ms vs Julia 3.7 ms  (5.3x)
  create_tile      JS  3.2 ms vs Julia 1.3 ms  (2.4x)
  get_mesh(30)     JS  1.2 ms vs Julia 0.18 ms (6.8x median)
  21-mesh sweep    JS 89.1 ms vs Julia 37.3 ms (2.4x)

+++ src/Martini.jl
@@ -106,16 +106,19 @@ function _update_errors!(tile::Tile)
         cx = mx + my - ay
         cy = my + ax - mx
 
-        interp = (terrain[ax, ay] + terrain[bx, by]) / 2
-        middle_err = abs(interp - terrain[mx, my])
-        errors[mx, my] = max(errors[mx, my], middle_err)
+        # Float64 intermediates match the JS reference: reading from a Float32Array
+        # promotes to a JS Number (Float64); Math.max(F32, F64) compares in F64 and
+        # truncates only on store back to the Float32Array. We mirror that exactly.
+        interp = (Float64(terrain[ax, ay]) + Float64(terrain[bx, by])) / 2
+        middle_err = abs(interp - Float64(terrain[mx, my]))
+        errors[mx, my] = Float32(max(Float64(errors[mx, my]), middle_err))
 
         if i < npt
-            errors[mx, my] = max(
-                errors[mx, my],
-                errors[(ax + cx) >> 1, (ay + cy) >> 1],
-                errors[(bx + cx) >> 1, (by + cy) >> 1],
-            )
+            errors[mx, my] = Float32(max(
+                Float64(errors[mx, my]),
+                Float64(errors[(ax + cx) >> 1, (ay + cy) >> 1]),
+                Float64(errors[(bx + cx) >> 1, (by + cy) >> 1]),
+            ))
         end
     end
     return tile
